Remove the usage of javah
with the jdk8 the javah command is deprecated and is removed at least
in the jdk10. The javac is now doing the work, the necessary changes
are added to the pom.xml.

To mark the parts which should be exposed into the header files, it is
important to use the java native annotation @Native.

Only the files which contains exposed information are generated.